+++ src/main/java/markdown/table/parser/MarkdownTableParser.kt
@@ -2,12 +2,12 @@ package markdown.table.parser
 
 import kotlinx.serialization.Decoder
 import kotlinx.serialization.KSerializer
-import kotlinx.serialization.NamedValueDecoder
-import kotlinx.serialization.SerialDescriptor
-import kotlinx.serialization.elementNames
 
 typealias DecoderFactory = (headerMap: Map<String, Int>, row: List<String>) -> Decoder
 
+/**
+ * マトリックスの行と列を入れ替える
+ */
 fun <T> List<List<T>>.transpose(): List<List<T>> {
     val numRows = size
     val numColumns = first().size
@@ -19,34 +19,9 @@ fun <T> List<List<T>>.transpose(): List<List<T>> {
     }
 }
 
-class MyNamedValueDecoder(
-    private val headerMap: Map<String, Int>,
-    private val row: List<String>
-) : NamedValueDecoder() {
-    override fun decodeTaggedString(tag: String): String {
-        val index = requireNotNull(headerMap[tag])
-        return row[index]
-    }
-
-    override fun decodeTaggedEnum(tag: String, enumDescription: SerialDescriptor): Int {
-        val index = requireNotNull(headerMap[tag])
-        val value = row[index]
-        return enumDescription.elementNames().indexOf(value)
-    }
-
-    override fun decodeTaggedBoolean(tag: String): Boolean {
-        val index = requireNotNull(headerMap[tag])
-        val value = row[index]
-        return value == "o"
-    }
-
-    override fun decodeTaggedNotNullMark(tag: String): Boolean {
-        val index = requireNotNull(headerMap[tag])
-        val value = row[index]
-        return value !in arrayOf("-", "−", "/")
-    }
-}
-
+/**
+ * マトリックスをParamリストに変換する
+ */
 fun <T> List<List<String>>.toParamList(serializer: KSerializer<T>, createDecoder: DecoderFactory): List<T> {
     val headerMap = first().mapIndexed { index, head -> head to index }.toMap()
 
@@ -56,6 +31,9 @@ fun <T> List<List<String>>.toParamList(serializer: KSerializer<T>, createDecoder
         .toList()
 }
 
+/**
+ * マークダウン形式の文字列をパースしてマトリックスを得る
+ */
 fun parseAsMarkdownTable(src: String): List<List<String>> {
     return src.split("\n")
         .asSequence()
